refactor: 포럼 채팅 컨트롤러, 서비스 query, command로 분리
