+++ src/backend/chat-service/src/main/java/capstone/chatservice/domain/forum/service/command/ForumMessageCommandService.java
@@ -1,0 +1,15 @@
+package capstone.chatservice.domain.forum.service.command;
+
+import capstone.chatservice.domain.forum.dto.ForumMessageDto;
+import capstone.chatservice.domain.forum.dto.request.ForumMessageCreateRequest;
+import capstone.chatservice.domain.forum.dto.request.ForumMessageDeleteRequest;
+import capstone.chatservice.domain.forum.dto.request.ForumMessageModifyRequest;
+
+public interface ForumMessageCommandService {
+
+    ForumMessageDto save(ForumMessageCreateRequest createRequest);
+
+    ForumMessageDto modify(ForumMessageModifyRequest modifyRequest);
+
+    ForumMessageDto delete(ForumMessageDeleteRequest deleteRequest);
+}
