+++ src/backend/chat-service/src/main/java/capstone/chatservice/domain/forum/service/query/ForumMessageQueryService.java
@@ -1,0 +1,11 @@
+package capstone.chatservice.domain.forum.service.query;
+
+import capstone.chatservice.domain.forum.dto.ForumMessageDto;
+import org.springframework.data.domain.Page;
+
+public interface ForumMessageQueryService {
+
+    Page<ForumMessageDto> getMessages(Long forumId, int page, int size);
+
+    Page<ForumMessageDto> getComments(Long parentId, int page, int size);
+}
